Moved shipping to checkout module
