+++ core/modules/checkout/index.ts
@@ -1,5 +1,6 @@
 import { checkoutModule } from './store/checkout'
 import { paymentModule } from './store/payment'
+import { shippingModule } from './store/shipping'
 
 import { VueStorefrontModule, VueStorefrontModuleConfig } from '@vue-storefront/module'
 
@@ -9,6 +10,7 @@ const moduleConfig: VueStorefrontModuleConfig = {
   key: KEY,
   store: { modules: [
     { key: 'payment', module: paymentModule },
+    { key: 'shipping', module: shippingModule },
     { key: 'checkout', module: checkoutModule },
   ] },
 }
