+++ core/modules/checkout/types/ShippingState.ts
