+++ core/store/modules/index.ts
@@ -1,10 +1,8 @@
-import shipping from './shipping'
 import ui from './ui-store'
 import sync from './sync'
 import promoted from './promoted-offers'
 
 export default {
-  shipping,
   ui,
   sync,
   promoted
